requirements.txt
ultralytics
opencv-python
torch
torchvision
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
numpy
scikit-image
omegaconf
pytorch-lightning
kornia
pyyaml
simple-lama-inpainting
Pillow
python-dotenv
